Rename file to remove base_
