+++ astropy/wcs/wcsapi/__init__.py
@@ -1,1 +1,1 @@
-from .base_low_level_api import *
+from .low_level_api import *
